requirements.txt
python-dotenv
pandas
numpy
flask
pymongo
certifi
scikit-learn
pymongo[srv]==3.11

# -e .